refactor db create mod slightly
